requirements.txt
flask==2.3.3
flask-cors==4.0.0
anthropic==0.18.1
openai==2.14.0
requests==2.31.0
litellm==1.52.14
numpy==1.26.4
chromadb==0.5.3
sentence-transformers==2.3.1
python-dotenv==1.0.0
google-generativeai==0.3.0
gunicorn==21.2.0
aiosqlite==0.21.0
